Add convenience method for configuring admin form
It is generally simpler to just use `AsyncFileField` where appropriate,
however in cases where it is not possible to adjust the model's field
type for whatever reason, this can be used in the `ModelAdmin`'s
`formfield_for_dbfield` method to achieve the same result.

E.g.
```
class CustomModelAdmin(admin.ModelAdmin):
    def formfield_for_dbfield(self, db_field, **kwargs):
        kwargs = make_resumable_admin_file_fields(db_field, ['file'], **kwargs)
        return super(AttachmentAdmin, self).formfield_for_dbfield(db_field, **kwargs)
```